requirements.txt
requests>=2.25.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy==1.7.1
seaborn>=0.11.0
networkx>=2.6.0
umap-learn>=0.5.0
tqdm>=4.60.0
aiohttp>=3.7.4
aiofiles>=0.7.0
flask>=2.0.0
werkzeug>=2.0.0
gunicorn>=20.1.0
